Correcting whitening testing
In previous notebook the whitening of the data was giving the lowest evaluate outcome.
However, some steps were missing in previous notebook. This notebook cleans up the whiten method.
But the lowest outcome is still about 7, which is not great. Further analysis in to the variables being used to predict to be done to prefect the model